src/requirements.txt
albumentations==1.3.0
opencv-python-headless==4.7.0.72
python-telegram-bot==13.15
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.1
tomli==2.0.1
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
matplotlib==3.7.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
